refactor: refactoring to MVC pattern
